tests/html/semantics/embedded-content-0/the-video-element/video_volume_volumechange.html

<!DOCTYPE html>
<html>
  <head>
    <title>Video Test: video_volume_volumechange</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-media-volume" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if video.volume is changed that expecting a simple event named volumechange is fired at the media element" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <video id="m">The user agent doesn't support media element.</video>
    <script type="text/javascript">
        var media = document.getElementById("m");
        var name = document.getElementsByName("assert")[0].content;
        var t = async_test(name);

        function startTest() {
            t.step(function() {
                assert_true(true, "volume changeed");
            });
            t.done();
        }

        media.addEventListener("volumechange", startTest, false);
        media.volume = media.defaultMuted? 1.0 : 0.0;
    </script>
  </body>
</html>
